Move code to service starter
