+++ service/lib/agama/dbus/storage_service.rb
@@ -25,6 +25,7 @@ require "agama/dbus/storage/iscsi"
 require "agama/dbus/storage/manager"
 require "agama/storage"
 require "y2storage/inhibitors"
+require "y2storage/storage_env"
 
 module Agama
   module DBus
@@ -55,7 +56,13 @@ module Agama
         # Inhibits various storage subsystem (udisk, systemd mounts, raid auto-assembly) that
         # interfere with the operation of yast-storage-ng and libstorage-ng.
         Y2Storage::Inhibitors.new.inhibit
-        manager.setup
+
+        # Underlying yast-storage-ng has own mechanism for proposing boot strategies.
+        # However, we don't always want to use BLS when it proposes so. Currently
+        # we want to use BLS only for Tumbleweed / Slowroll
+        prohibit_bls_boot if !config.boot_strategy&.casecmp("BLS")
+
+        check_multipath
         export
       end
 
@@ -83,6 +90,30 @@ module Agama
 
       # @return [Logger]
       attr_reader :logger
+
+      def prohibit_bls_boot
+        ENV["YAST_NO_BLS_BOOT"] = "1"
+        # avoiding problems with cached values
+        Y2Storage::StorageEnv.instance.reset_cache
+      end
+
+      MULTIPATH_CONFIG = "/etc/multipath.conf"
+      private_constant :MULTIPATH_CONFIG
+
+      # Checks if all requirement for multipath probing is correct and if not then log it.
+      def check_multipath
+        # check if kernel module is loaded
+        mods = `lsmod`.lines.grep(/dm_multipath/)
+        logger.warn("dm_multipath modules is not loaded") if mods.empty?
+
+        binary = system("which multipath")
+        if binary
+          conf = `multipath -t`.lines.grep(/find_multipaths "smart"/)
+          logger.warn("multipath: find_multipaths is not set to 'smart'") if conf.empty?
+        else
+          logger.warn("multipath is not installed.")
+        end
+      end
 
       # @return [::DBus::ObjectServer]
       def service
